+++ config/auth.js
@@ -99,5 +99,14 @@ const get = (key) => {
   return config ? config : {};
 }
 
+// for now admin users can login in with Password (Local) and a Url emailed
+exports.adminTypes = types.filter((type) => ['Local', 'Url'].includes(type.key)).map((type) => {
+  // for admin types we add /admin to the login url
+  return Object.assign({}, type, {
+    loginUrl: type.loginUrl + '/admin'
+  })
+})
+
+
 exports.types = types;
 exports.get = get;