+++ routes/routes.js
@@ -9,9 +9,7 @@ const userController = require('../controllers/user/user');
 const authChoose = require('../controllers/auth/choose');
 const authUrl = require('../controllers/auth/url');
 const authPhonenumber = require('../controllers/auth/phonenumber');
-const authAdminUrl = require('../controllers/auth/adminUrl');
 const authForgot = require('../controllers/auth/forgot');
-const authDigiD = require('../controllers/auth/digid');
 const authAnonymous = require('../controllers/auth/anonymous');
 const authLocal = require('../controllers/auth/local');
 const authCode = require('../controllers/auth/code');
@@ -25,6 +23,9 @@ const bruteForce = require('../middleware/bruteForce');
 const authMw = require('../middleware/auth');
 const passwordResetMw = require('../middleware/passwordReset');
 const logMw = require('../middleware/log');
+
+//UTILS
+const getClientIdFromRequest = require('../utils/getClientIdFromRequest');
 
 //MODELS
 const ExternalCsrfToken = require('../models').ExternalCsrfToken;
@@ -57,10 +58,16 @@ const smsCodeBruteForce = bruteForce.user.getMiddleware({
     }
 });
 
-const emailUrlBruteForce = bruteForce.user.getMiddleware({
+// prevent too many attempts for the same email per clientId (if applicable)
+const emailUrlBruteForce = bruteForce.userVeryRestricted.getMiddleware({
     key: function (req, res, next) {
-        // prevent too many attempts for the same email
-        next(req.body.email);
+        const clientId = getClientIdFromRequest(req);
+        
+        if (clientId) {
+            return next(`${clientId}-${req.body.email}`);
+        }
+        
+        return next(`${req.body.email}`);
     }
 });
 
@@ -194,7 +201,7 @@ module.exports = function (app) {
      * checks if one or more options of authentications is available
      * and either shows choice or redirects if one option
      */
-    app.get('/login', clientMw.withOne, authChoose.index);
+    app.get('/login/:priviligedRoute?', clientMw.withOne, authChoose.index);
 
     /**
      * Shared middleware for all auth routes, adding client and bruteforce
@@ -205,44 +212,42 @@ module.exports = function (app) {
      * Login & register with local login
      */
     //shared middleware
-    app.use('/auth/local', [clientMw.setAuthType('Local'), clientMw.validate, csrfProtection, addCsrfGlobal]);
-
+    const localSharedMW = [clientMw.setAuthType('Local'),  csrfProtection, addCsrfGlobal]
     //routes
-    app.get('/auth/local/login', authLocal.login);
-    app.post('/auth/local/login', csrfProtection, loginBruteForce, authMw.validateLogin, authLocal.postLogin);
-    app.get('/auth/local/register',  csrfProtection, addCsrfGlobal, authLocal.register);
-    app.post('/auth/local/register',  csrfProtection, addCsrfGlobal, userMw.validateUser, userMw.validateUniqueEmail, authLocal.postRegister);
+    app.get('/auth/local/login/:priviligedRoute?', [...localSharedMW], clientMw.validate, authLocal.login);
+    app.post('/auth/local/login/:priviligedRoute?',  [...localSharedMW],  csrfProtection, loginBruteForce, authMw.validateLogin, authLocal.postLogin);
+    app.get('/auth/local/register',  [...localSharedMW],  csrfProtection, addCsrfGlobal, authLocal.register);
+    app.post('/auth/local/register',  [...localSharedMW],  csrfProtection, addCsrfGlobal, userMw.validateUser, userMw.validateUniqueEmail, authLocal.postRegister);
 
     /**
      * Deal with forgot password
      */
-    app.get('/auth/local/forgot', authForgot.forgot);
-    app.post('/auth/local/forgot', authForgot.postForgot);
-    app.get('/auth/local/reset', passwordResetMw.validate, authForgot.reset);
-    app.post('/auth/local/reset', passwordResetMw.validate, userMw.validatePassword, authForgot.postReset);
+    app.get('/auth/local/forgot',  [...localSharedMW], authForgot.forgot);
+    app.post('/auth/local/forgot', [...localSharedMW], authForgot.postForgot);
+    app.get('/auth/local/reset',  [...localSharedMW], passwordResetMw.validate, authForgot.reset);
+    app.post('/auth/local/reset', [...localSharedMW], passwordResetMw.validate, userMw.validatePassword, authForgot.postReset);
 
     /**
      * Auth routes for URL login
      */
     // shared middleware
-    app.use('/auth/url', [clientMw.setAuthType('Url'), clientMw.validate]);
+    ///app.use('/auth/url', );
 
     // routes
-    app.get('/auth/url/login', csrfProtection, addCsrfGlobal, authUrl.login);
-    app.get('/auth/url/confirmation', csrfProtection, addCsrfGlobal, authUrl.confirmation);
-    app.post('/auth/url/login', csrfProtection, emailUrlBruteForce, authUrl.postLogin);
-    app.get('/auth/url/authenticate', csrfProtection, addCsrfGlobal, authUrl.authenticate);
-    app.post('/auth/url/authenticate', csrfProtection, emailUrlBruteForce, authUrl.postAuthenticate);
+    app.get('/auth/url/login/:priviligedRoute?', clientMw.setAuthType('Url'), clientMw.validate, csrfProtection, addCsrfGlobal, authUrl.login);
+    app.get('/auth/url/confirmation', clientMw.setAuthType('Url'),  csrfProtection, addCsrfGlobal, authUrl.confirmation);
+    app.post('/auth/url/login/:priviligedRoute?', clientMw.setAuthType('Url'), csrfProtection, emailUrlBruteForce, authUrl.postLogin);
+    app.get('/auth/url/authenticate', clientMw.setAuthType('Url'),  csrfProtection, addCsrfGlobal, authUrl.authenticate);
+    app.post('/auth/url/authenticate', clientMw.setAuthType('Url'), csrfProtection, emailUrlBruteForce, authUrl.postAuthenticate);
 
 
-    // Admin login routes
-    app.use('/auth/admin', [csrfProtection, addCsrfGlobal]);
+    // admin login routes redirect to normal login but with priviliged params
+    app.get('/auth/admin/login', [csrfProtection, addCsrfGlobal], (req, res, next) => {
+        const queryIndex = req.originalUrl.indexOf('?');
+        const queryString = (queryIndex>=0) ? req.originalUrl.slice(queryIndex) : '';
 
-    app.get('/auth/admin/login', authUrl.login);
-    app.get('/auth/admin/confirmation', authAdminUrl.confirmation);
-    app.post('/auth/admin/login', emailUrlBruteForce, authAdminUrl.postLogin);
-    app.get('/auth/admin/authenticate', authUrl.authenticate);
-    app.post('/auth/admin/authenticate', emailUrlBruteForce, authAdminUrl.postAuthenticate);
+        res.redirect('/login/admin' + queryString);
+    });
 
     /**
      * Auth routes for Anonymous login
@@ -293,7 +298,7 @@ module.exports = function (app) {
     app.post('/password', clientMw.withOne, authMw.check, csrfProtection, addCsrfGlobal, userMw.validatePassword, userController.postAccount);
 
     app.use('/auth/required-fields', [authMw.check, clientMw.withOne]);
-    app.get('/auth/required-fields', clientMw.withOne, clientMw.checkIfEmailRequired, authRequiredFields.index);
+    app.get('/auth/required-fields', clientMw.withOne, csrfProtection, addCsrfGlobal, clientMw.checkIfEmailRequired, authRequiredFields.index);
     app.post('/auth/required-fields', clientMw.withOne, csrfProtection, addCsrfGlobal, authRequiredFields.post);
 
     app.use('/auth/two-factor', [authMw.check, clientMw.withOne]);
@@ -304,14 +309,15 @@ module.exports = function (app) {
 
     app.use('/dialog', [bruteForce.global.prevent]);
 
-    app.get('/dialog/authorize', clientMw.withOne, authMw.check, userMw.withRoleForClient, clientMw.checkRequiredUserFields, clientMw.check2FA, clientMw.checkUniqueCodeAuth((req, res) => {
+    app.get('/dialog/authorize', clientMw.withOne, authMw.check, userMw.withRoleForClient, clientMw.checkRequiredUserFields, clientMw.check2FA, clientMw.checkPhonenumberAuth(), clientMw.checkUniqueCodeAuth((req, res) => {
         return res.redirect('/login?clientId=' + req.query.client_id);
     }), oauth2Controller.authorization);
-    app.post('/dialog/authorize/decision', clientMw.withOne, userMw.withRoleForClient, clientMw.checkUniqueCodeAuth(),clientMw.check2FA, bruteForce.global.prevent, oauth2Controller.decision);
+
+    app.post('/dialog/authorize/decision', clientMw.withOne, userMw.withRoleForClient, clientMw.checkPhonenumberAuth(), clientMw.checkUniqueCodeAuth(),clientMw.check2FA, bruteForce.global.prevent, oauth2Controller.decision);
     app.post('/oauth/token', oauth2Controller.token);
     app.get('/oauth/token', oauth2Controller.token);
 
-    app.get('/api/userinfo', passport.authenticate('bearer', {session: false}), clientMw.withOne, userMw.withRoleForClient, clientMw.checkUniqueCodeAuth(), userController.info);
+    app.get('/api/userinfo', passport.authenticate('bearer', {session: false}), clientMw.withOne, userMw.withRoleForClient, clientMw.checkIfAccessTokenBelongToCurrentClient, clientMw.checkPhonenumberAuth(), clientMw.checkUniqueCodeAuth(), userController.info);
 
     // Mimicking google's token info endpoint from
     // https://developers.google.com/accounts/docs/OAuth2UserAgent#validatetoken
@@ -329,8 +335,22 @@ module.exports = function (app) {
     });
 
     // Handle 500
-    app.use(function (err, req, res, next) {
+    app.use(async function (err, req, res, next) {
         console.log('===> err', err);
+        // een deserialize error betekent een data fout; daar hoef je een gebruiker niet mee te belasten
+        if (err && err.message && err.message.match(/^Error in deserializeUser/)) {
+          console.log(err); // do log for debugging
+          await req.session.destroy();
+          let querystring = '?'
+          if (req.query.clientId) querystring += `&clientId=${req.query.clientId}`;
+          if (req.query.client_id) querystring += `&client_id=${req.query.client_id}`;
+          if (req.query.redirectUrl) querystring += `&redirectUrl=${encodeURIComponent(req.query.redirectUrl)}`;
+          if (req.query.redirect_uri) querystring += `&redirect_uri=${encodeURIComponent(req.query.redirect_uri)}`;
+          if (req.query.token) querystring += `&token=${req.query.token}`;
+          if (req.query.access_token) querystring += `&access_token=${req.query.access_token}`;
+          return res.redirect('/logout'+querystring);
+        }
         res.status(500).render('errors/500');
     });
+
 }