+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## v0.40.0
+Update version number
+
+## 0.19.0
+* Make the email url brute force more restrictive
+
+## 0.18.0
+* Remove password from return hashed in GET 
+
+## 0.17.0
+* Refactor logic so admin can login with both Password and Login via E-mail
+
 ## v0.16.0 (2020-07-20)
 * Add external CSRF token